requirement.txt
python-decouple>=3.8
python-dotenv>=1.0.0
SQLAlchemy>=2.0.18
uvicorn>=0.22.0
fastapi>=0.99.1
PyMySQL>=1.1.0
pytest>=7.4.0
httpx>=0.24.1

python-dateutil>=2.8.2
schedule>=1.2.0

